# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Football Prediction API
  description: API for football match prediction application
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MatchSchedule:
      type: object
      properties:
        id:
          type: integer
        team1_id:
          type: integer
        team2_id:
          type: integer
        date:
          type: string
          format: date
        time:
          type: string
          format: time
        skor1:
          type: integer
        skor2:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        team1:
          $ref: '#/components/schemas/Team'
        team2:
          $ref: '#/components/schemas/Team'
...
paths:
...
  /match-schedules:
    get:
      tags:
        - Match Schedules
      summary: Get all match schedules
      responses:
        '200':
          description: List of match schedules
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MatchSchedule'

    post:
      tags:
        - Match Schedules
      summary: Create a new match schedule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - team1_id
                - team2_id
                - date
                - time
              properties:
                team1_id:
                  type: integer
                team2_id:
                  type: integer
                date:
                  type: string
                  format: date
                time:
                  type: string
                  format: time
      responses:
        '201':
          description: Match schedule created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/MatchSchedule'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /match-schedules/{id}:
    get:
      tags:
        - Match Schedules
      summary: Get match schedule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Match schedule details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/MatchSchedule'
        '404':
          description: Match schedule not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string

    put:
      tags:
        - Match Schedules
      summary: Update match schedule
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                team1_id:
                  type: integer
                team2_id:
                  type: integer
                date:
                  type: string
                  format: date
                time:
                  type: string
                  format: time
                skor1:
                  type: integer
                skor2:
                  type: integer
      responses:
        '200':
          description: Match schedule updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/MatchSchedule'
        '404':
          description: Match schedule not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string

  # News Routes
...
security:
  - bearerAuth: []